[halium-wrappers] Added new utility 'waitforservice'
Logic of this is trivial: wait until at least one of the supplied
system properties is set to 'running'.
This is especially useful when trying to determine whether a
service started by the Android init is now running.
For example,
    waitforservice \
        init.svc.vendor.hwcomposer-2-1 \
        init.svc.vendor.hwcomposer-2-2 \
        init.svc.vendor.hwcomposer-2-*
Will return once at least one of the supplied properties compares
to 'running', i.e. Android init reported that as running.
Wildcards are supported, pattern matching is being done with
fnmatch().

This is accomplished via bionic's __system_property_wait_any()
function, which is deprecated (but still available at least up
to Android R).
There is no way to set a timeout yet.

Signed-off-by: Eugenio Paolantonio (g7) <me@medesimo.eu>